update: format code accroding to eslint config
