+++ packages/plugin-cosmos/eslint.config.mjs
@@ -25,7 +25,8 @@ export default [
         { blankLine: 'any', prev: ['const', 'let', 'var'], next: ['const', 'let', 'var'] },
         { blankLine: 'always', prev: '*', next: 'block-like' }
       ],
-      "no-unused-vars": [
+      "no-unused-vars": "off",
+      "@typescript-eslint/no-unused-vars": [
         "error",
         {
           argsIgnorePattern: "^_",
