+++ packages/plugin-cosmos/src/tests/fee-estimator.test.ts
@@ -3,92 +3,93 @@ import { FeeEstimator } from "../services/fee-estimator";
 import { SigningCosmWasmClient } from "@cosmjs/cosmwasm-stargate";
 
 vi.mock("@cosmjs/cosmwasm-stargate", () => ({
-    SigningCosmWasmClient: {
-        simulate: vi.fn(),
-    },
+  SigningCosmWasmClient: {
+    simulate: vi.fn(),
+  },
 }));
 
 describe("FeeEstimator", () => {
-    let mockSigningCosmWasmClient: SigningCosmWasmClient;
-    let feeEstimator: FeeEstimator;
+  let mockSigningCosmWasmClient: SigningCosmWasmClient;
+  let feeEstimator: FeeEstimator;
 
-    beforeEach(() => {
-        mockSigningCosmWasmClient = {
-            simulate: vi.fn(),
-        } as unknown as SigningCosmWasmClient;
+  beforeEach(() => {
+    mockSigningCosmWasmClient = {
+      simulate: vi.fn(),
+    } as unknown as SigningCosmWasmClient;
 
-        feeEstimator = new FeeEstimator(mockSigningCosmWasmClient);
+    feeEstimator = new FeeEstimator(mockSigningCosmWasmClient);
 
-        vi.clearAllMocks();
-    });
+    vi.clearAllMocks();
+  });
 
-    it("should estimate gas for sending tokens successfully", async () => {
-        const mockGasEstimation = 200000;
-        // @ts-ignore
-        (mockSigningCosmWasmClient.simulate as vi.Mock).mockResolvedValue(
-            mockGasEstimation
-        );
+  it("should estimate gas for sending tokens successfully", async () => {
+    const mockGasEstimation = 200000;
 
-        const senderAddress = "cosmos1senderaddress";
-        const recipientAddress = "cosmos1recipientaddress";
-        const amount = [{ denom: "uatom", amount: "1000000" }];
-        const memo = "Test memo";
+    // @ts-expect-error -- ...
+    (mockSigningCosmWasmClient.simulate as vi.Mock).mockResolvedValue(
+      mockGasEstimation
+    );
 
-        const estimatedGas = await feeEstimator.estimateGasForSendTokens(
-            senderAddress,
-            recipientAddress,
-            amount,
-            memo
-        );
+    const senderAddress = "cosmos1senderaddress";
+    const recipientAddress = "cosmos1recipientaddress";
+    const amount = [{ denom: "uatom", amount: "1000000" }];
+    const memo = "Test memo";
 
-        expect(estimatedGas).toBe(mockGasEstimation);
-        expect(mockSigningCosmWasmClient.simulate).toHaveBeenCalledWith(
-            senderAddress,
-            [
-                {
-                    typeUrl: "/cosmos.bank.v1beta1.MsgSend",
-                    value: {
-                        fromAddress: senderAddress,
-                        toAddress: recipientAddress,
-                        amount: [...amount],
-                    },
-                },
-            ],
-            memo
-        );
-    });
+    const estimatedGas = await feeEstimator.estimateGasForSendTokens(
+      senderAddress,
+      recipientAddress,
+      amount,
+      memo
+    );
 
-    it("should throw an error if gas estimation fails", async () => {
-        // @ts-ignore
-        (mockSigningCosmWasmClient.simulate as vi.Mock).mockRejectedValue(
-            new Error("Gas estimation failed")
-        );
+    expect(estimatedGas).toBe(mockGasEstimation);
+    expect(mockSigningCosmWasmClient.simulate).toHaveBeenCalledWith(
+      senderAddress,
+      [
+        {
+          typeUrl: "/cosmos.bank.v1beta1.MsgSend",
+          value: {
+            fromAddress: senderAddress,
+            toAddress: recipientAddress,
+            amount: [...amount],
+          },
+        },
+      ],
+      memo
+    );
+  });
 
-        const senderAddress = "cosmos1senderaddress";
-        const recipientAddress = "cosmos1recipientaddress";
-        const amount = [{ denom: "uatom", amount: "1000000" }];
+  it("should throw an error if gas estimation fails", async () => {
+    // @ts-expect-error -- ...
+    (mockSigningCosmWasmClient.simulate as vi.Mock).mockRejectedValue(
+      new Error("Gas estimation failed")
+    );
 
-        await expect(
-            feeEstimator.estimateGasForSendTokens(
-                senderAddress,
-                recipientAddress,
-                amount
-            )
-        ).rejects.toThrow("Gas estimation failed");
+    const senderAddress = "cosmos1senderaddress";
+    const recipientAddress = "cosmos1recipientaddress";
+    const amount = [{ denom: "uatom", amount: "1000000" }];
 
-        expect(mockSigningCosmWasmClient.simulate).toHaveBeenCalledWith(
-            senderAddress,
-            [
-                {
-                    typeUrl: "/cosmos.bank.v1beta1.MsgSend",
-                    value: {
-                        fromAddress: senderAddress,
-                        toAddress: recipientAddress,
-                        amount: [...amount],
-                    },
-                },
-            ],
-            ""
-        );
-    });
+    await expect(
+      feeEstimator.estimateGasForSendTokens(
+        senderAddress,
+        recipientAddress,
+        amount
+      )
+    ).rejects.toThrow("Gas estimation failed");
+
+    expect(mockSigningCosmWasmClient.simulate).toHaveBeenCalledWith(
+      senderAddress,
+      [
+        {
+          typeUrl: "/cosmos.bank.v1beta1.MsgSend",
+          value: {
+            fromAddress: senderAddress,
+            toAddress: recipientAddress,
+            amount: [...amount],
+          },
+        },
+      ],
+      ""
+    );
+  });
 });
